requirements.txt
# python 3.13
pylint==3.2.5
pytest==8.2.2
pytest-cov==5.0.0
ipython==8.26.0
black==24.4.2
slack_bolt
setuptools==70.3.0
requests~=2.32.3
python-dotenv~=1.0.1
flask~=3.0.3
pyyaml~=6.0.1


